+++ src/main/java/com/example/ootd/domain/clothes/repository/AttributeRepository.java
@@ -1,6 +1,7 @@
 package com.example.ootd.domain.clothes.repository;
 
 import com.example.ootd.domain.clothes.entity.Attribute;
+import com.example.ootd.domain.clothes.repository.custom.CustomAttributeRepository;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
