+++ src/main/java/com/example/ootd/domain/clothes/repository/ClothesRepository.java
@@ -1,6 +1,7 @@
 package com.example.ootd.domain.clothes.repository;
 
 import com.example.ootd.domain.clothes.entity.Clothes;
+import com.example.ootd.domain.clothes.repository.custom.CustomClothesRepository;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
