+++ src/main/java/com/example/ootd/domain/clothes/repository/custom/CustomAttributeRepository.java
@@ -1,4 +1,4 @@
-package com.example.ootd.domain.clothes.repository;
+package com.example.ootd.domain.clothes.repository.custom;
 
 import com.example.ootd.domain.clothes.dto.request.ClothesAttributeSearchCondition;
 import com.example.ootd.domain.clothes.entity.Attribute;
