+++ src/main/java/com/example/ootd/domain/feed/repository/FeedRepository.java
@@ -1,9 +1,10 @@
 package com.example.ootd.domain.feed.repository;
 
 import com.example.ootd.domain.feed.entity.Feed;
+import com.example.ootd.domain.feed.repository.custom.CustomFeedRepository;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FeedRepository extends JpaRepository<Feed, UUID> {
+public interface FeedRepository extends JpaRepository<Feed, UUID>, CustomFeedRepository {
 
 }
