--- src/main/java/com/example/ootd/domain/feed/repository/impl/CustomFeedRepositoryImpl.java
@@ -1,11 +1,0 @@
-package com.example.ootd.domain.feed.repository.impl;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Repository;
-
-@Repository
-@RequiredArgsConstructor
-public class CustomFeedRepositoryImpl {
-
-
-}
